requirements.txt
auth==0.5.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
eventlet==0.39.0
falcon==4.0.2
Flask==3.1.0
Flask-Cors==5.0.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mongoengine==0.29.1
packaging==24.2
PyJWT==2.10.1
pymongo==4.11.1
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
